Factor classes into separate module file
